Create README.md
I hold a degree (BAC+5) in Economics, with a focus on statistical and econometric methods. Since March 2021, I have been working as a Junior Analyst-Economist at the economic research firm "Congo Challenge". In addition to my education as economist, I took online training in Data Science with the R language. This training gave me the opportunity to sharpen my skills in Statistics and Data Analysis, and to master the R programming language for both Data Science and Machine Learning. In this perspective, I had the opportunity to participate in a social housing survey in Kinshasa where I was statistics assistant. This experience allowed me to reconcile my theoretical background in Applied Statistics and Data Science with the practice on real databases. Finally, I was able to build an ML algorithm that would allow the Foundation where I am volunteer to predict the eligibility of scholarship candidates based on their profiles (age, academic performance, language level, etc.).